backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
drf-spectacular
psycopg[binary]
Pillow
numpy
pandas
scikit-learn
requests
sentence-transformers
python-dotenv
opencv-python-headless
ultralytics
pytesseract